Add files via upload
The EDA is complete. The pre-processing phase has been completed. Several models such as RandomForestClassifier, SVC, and LinearSVC have been trained. Model overfitting has been completed, and fine-tuning is underway for individual models. Ensemble learning of the various tested models has begun, awaiting the results of the fine-tuning. The results are very satisfactory; all that remains is to try to maximize them.